week-05-tasks/requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
langchain-pinecone
langchain-huggingface
pinecone-client
sentence-transformers
PyMuPDF
python-docx
